+++ docs/RELEASE-NOTES.md
@@ -1,5 +1,18 @@
 # Release Notes
 
+## 1.1.0 - 24 February 2021
+
+Breaking changes:
+- (None)
+
+New features:
+- Added `ServiceCollectionExtensions.GetConfiguration`.
+- Added `ServiceCollectionExtensions.GetConfigurationDescriptors`.
+- Added `ServiceDescriptorExtensions.GetConfigurationInstance`.
+
+Bug fixes / internal changes:
+- (None)
+
 ## 1.0.4 - 20 January 2021
 
 Breaking changes: